requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.3
openai==1.3.5
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
httpx==0.25.1
